+++ test/unified_format.py
@@ -985,9 +985,9 @@ class UnifiedSpecTestMixinV1(IntegrationTest):
                       "session %s" % (spec['session'],))
 
         client = single_client('%s:%s' % session._pinned_address)
+        self.addCleanup(client.close)
         self.__set_fail_point(
             client=client, command_args=spec['failPoint'])
-        self.addCleanup(client.close)
 
     def _testOperation_assertSessionTransactionState(self, spec):
         session = self.entity_map[spec['session']]